requirements.txt
python-bitvavo-api==1.3.3
python-dotenv==1.0.1
pandas==2.2.2
pandas-ta==0.3.14b0
requests==2.31.0
APScheduler==3.10.4
paho-mqtt==1.6.1
